Lesson 5: Add new font
Add the Roboto Condensed font.
Add the Home route.
