+++ src/components/category-item/category-item.component.jsx
@@ -14,7 +14,7 @@ const CategoryItem = ({ category }) => {
       />
       <div className="category-body-container">
         <h2>{title}</h2>
-        <p>Shop Now</p>
+        <p>Shop Now Here</p>
       </div>
     </div>
   );
